requirements.txt
Twisted==16.6
Scrapy==1.2.3
django_autocomplete_light==3.2.1
Django==1.10.3
django_filter==1.0.2
nltk==3.2.2
psycopg2==2.6.2
Pillow
pypiwin32